Bug with getplayertank found and fixed. Bug with beginturn in AIoppentents found
GetPlayerTank was trying to index directly from a given player number ,
in bullet class it was being given a player number from 1 to 8 and
in AIoppentent class it was being given a player number from 0 to 7.
Changed both references to give a player number from 1 to 8 and GetPlayerTank
to handle changing the playernum to a indexable number.

BeginTurn was using some odd min and max values for firing angles in odd cases.
Changed the code in CalMinAndMaxAngle to produce more constant values. As it
could produce some weird min that would be larger than the max when down to 1v1 AI
tanks.