Added a new AttackEffect called HitText . Improved Hitbox calculation again , still not on point.
The new HitText class will show text on the screen for a short time
and can be implemented the same as a explosion would be into the battefeild
via the Battle class.

Improved the Hitbox calculation by including the tank's model witdth and height
into the distance for determining if it is inside a explosions radius. Still having
issues with it matching the visualization of explosions on the right side of tanks.